Allow for uint and string ASN
Use an uint ASN when doing the lookup. This will allow me to
attach easily with ooni/probe-cli.

Use a string inside the session, where it's probably what we want
since mobile apps expect a string there.

While there, reckon that, from the implementation PoV, the lookup
of the network name should be dominated by the ASN lookup.

The rational thing to do would be to do both at the same time, yet
I think the task API is such that separate is better.