mmdblookup: do not return empty strings
We already have reasonable defaults. Make sure we're not going to
return empty strings when we instead could keep the defaults.

Note that it may sometime happen that MMDB cannot return a location
for an IP address, e.g., when it's in the cloud. In such cases, it
is probably reasonable to keep "ZZ to mean "we don't know".