Start making the code more resilient
There are some network errors that we don't want to stop OONI from
running. We want to use a context and possibly interrupt it using
the context, otherwise, we want it to continue running.

This is not the perfect solution but it's a good start and it does
package the functionality in its own package, such that we can later
just go there and rollout better code.