Make sure proxying and custom CA work for real
Specifically, I've tested this in a Debian VM with no CA bundle
and Tor running. And I've verified that all that needed to be
proxied was proxied through Tor, and what needed not to be proxied
was not proxied instead. This set of changes requires some judo
around private variables of the mlabns package. Also, we need to
make the session.New process aware of an explicit proxy.