measurementkit: allow to select task's verbosity
This diff allows the caller to choose the verbosity of individual
measurementkit tasks rather than using "INFO" in any case.

The default is "WARNING" now. Which is consistent with what @hellais
said that he wants to control the output in the common case.

I need this functionality to understand why hhfm and hirl do not
work correctly with probe-cli. I'll keep this also after that since
in general this seems an useful feature to me.