API CHANGE: never redact probeIP
Experience integrating with probe-cli suggests that the code is
much easier to review and understand if we strip all the possibly
sensitive values like probeIP in the probe-cli bit of code that
deals with honouring such settings.

Deferring this job to probe-engine makes it more complex to reason
about what happens. Plus, a savvy probe-cli programmer is in any
case expected to perform this check in probe-cli.