collector: implement scrubbing safety net
This diff introduces code to redact IP, ASN, and CC directly
in the engine, rather than delegating doing that.

When you initialise a Session, it's your responsibility to
change the default privacy settings.

The names of the collector.PrivacySettings structure fields
are the same names used by ooni/probe-cli.

After a measurement, the settings configured in the session are
honoured with respect to redacting IP, ASN, and CC.

We take an extra step of redacting the whole test keys
string, _if_ it contains the IP.

This is an extra safety net to make sure that we are really
scrubbing the body. The nettest should do that.

This commit is part of #28.