Implement TCP connect and HTTP OONI templates
Notwithstanding some limitations, this implementation is already
close enough to what we'll need in production to be able to produce
a nearly-compliant Telegram JSON report.

The two main limitations is that we could take advantage more of
httpx/httpracex to collect extra low level information and the fact
that some keys that are supposed to be emitted as `null` are in
fact actually emitted as empty strings instead.

We will work through these limitations with future commits.