Initial implementation of DoH experiment
I'm printing on stdout the stream of events for now. Plan is to
mere this and have machinery. Then, do work encompassing all the
experiments using netx to produce common data format.

Spec for now is very vague with respect to the data format
specifically for this reason. Aim is to revise the spec once
we have written common netx processing code.