Write more functional tests for psiphon
Similar to what I did for Telegram and DASH, except that here there is still
some integration testing because it's more practical.

The real integration test has been moved to toplevel, again exactly
like I have done for Telegram and DASH.

While working on unit tests, I also stumbled upon a very annoying issue in
the design of netx porcelain: https://github.com/ooni/netx/issues/141.

Writing this diff has significantly changed the structure of the psiphon
test, which now is in a shape in which testing is easier to do.