psiphon: refine the output data model (#144)
After looking into the data model of the previous version, I have
determined it's not the case to seek for 100% backwards compat.

I have added all the ooni/probe-legacy psiphon experiment keys that
are still easy and meaningful. Here's what has changed:

1. the name of the temporary file is not relevant anymore as are
not relevant the exit code and the stdout/stderr

2. psiphon_installed does not make any sense

3. success could still be here, but TBH I belive it's much more
informative to have a nullable failure string

4. body is not here because we already have it in the HTTP requests

5. default_configuration is a key for which I cannot totally
understand what could be the use in this new scenario, but we
may probably want to discuss that?

6. I've chosen to use the HTTP data format of Web Connectivity rather
than the older one of HTTPRequests, because this is what we are documenting
now and because this is what the pipeline mostly process

7. in the example in the spec the agent being used does not implement
redirection, yet it's annoying to disable redirection here in Go, hence
I am advertising an agent using redirection

I should also have a chat with @FedericoCeratto se we can sync up
again with respect to the output data format of Psiphon, since I was
wrong when I told him the data format was already final.