telegram/test: correctly define success accessing https (#174)
This should close #142. Because maps have no stable sorting in Go, we
sometimes pick the success case first. However, such case is incomplete
because it does not set the status code as 200. Therefore, we see an
error telling us that the server returned a bad status code. Initialising
the variable to 200 should be enough to fix the bug.