session: introduce generic key-value store
By default the key-value store is in memory. We will be able to
easily override it for ooni/probe-cli. We should be able to do the
same also on mobile, because it is a Go interface.

This is part of #164, but we still have a bunch of work to do
before we can call this strand of work closed.