model: make kvstore easy to use with java and sqlite3
While in Go strings can be binary, this is not the case with java
and sqlite3. So, be very clear that the values may be binary while
the keys are expected to always be strings.

This addresses concerns around the resulting API when binding with
Go mobile, or when using sqlite3, where the key-value store could be
fail because the values are not UTF-8 (or UTF-16) strings.