engine: use rogpeppe/go-internal/lockedfile for kvstore
This patch changes the way in which we implement the kvstore used
to keep track of orchestra, to use rogpeppe/go-internal/lockedfile
rather than a database-like kvstore.

For our use case, atomic file read and write is adequate and we
don't need to add additional complexity.

Sadly, I didn't notice go-internal/lockedfile before.

Because we're not using a completely different kvstore, I have
changed to directory name to be `kvstore2`.