oonitemplates: implement OBFS4Connect template
There will later be opportunity to improve the code and also
consider extending to use, e.g., meeklite as well.

There will also be opportunity for using a better integration
testing obfs4proxy instance than the one we're using.

For now, we've the possibility of connecting using obfs4.

This work is part of #90.