oonitemplates.go: sanitise conn_id
We use port numbers in netx because this is a way to make sure
that we can bind together HTTP and net traces.

We use the sign to indicate TCP or UDP.

There may be value in not scrambling the ports inside netx (I am
not sure TBH), but surely it's better to avoid doing that here.

This is also part of #12.