Allow compiling with CGO_ENABLED=1 and no measurement-kit (#344)
I want to build an `.aar` file that only includes experiments written in Go such that I can link such `.aar` file into the OONI app along with measurement-kit/android-libs. To this end, I need to be able to compile this repository with CGO enabled, to cross compile for Android, but without linking with MK.

This work is part of https://github.com/ooni/probe-engine/issues/339.