add build-android.bash: script to build AAR
We create a new directory called MOBILE. This directory will
contain two subdirs, both `gitignored`. We will put output
packages inside the `dist` subdir. We will use the `gopath`
subdir as the temporary gopath.

The `go.sum` file changes slightly because we are using
content from the `x/mobile` repository. We are however not
compiling and building `gomobile` in a module enabled env
to avoiding tainting `go.sum` and `go.mod` too much. We are
in fact using a temporary gopath precisely for building
`gomobile` outside of any module environment.

This is part of https://github.com/measurement-kit/measurement-kit/issues/1913
and specifically it contains the strategy to not need MK for
running OONI on Android platforms. Of course, we are not at the
point where we can get rid of MK completely yet, but already
we have an AAR that implements some of our experiments.