orchestra: register desktop clients with AS0 and ZZ (#375)
When registering, we don't need to lookup the country and the ASN. Desktop
clients only use orchestra to fetch targets for tor and psiphon. Also there
is no push notifications on desktop. Thus, it seems to me there is really
no point in doing an extra CC and ASN lookup in this code. We may need to add
a special case for mobile later, but this seems to me the best default.

This is part of the QA for https://github.com/ooni/probe/issues/1028.