ExperimentMeasurer describes how an experiment measures
This diff modifies the way in which we tell the generic experiment
code about the specific experiment code.

This change is motived by these three issues:

1. https://github.com/ooni/probe-engine/issues/184

2. https://github.com/ooni/probe-engine/issues/183

3. https://github.com/ooni/probe-engine/issues/302

Very briefly, I have discovered that I need #184 and #185 in
order to do #302 in a clean way. In fact, to properly call #302
finished, I need to add to the session a Close method. This
method will guarantee that we don't leak open TCP connections
into the HTTP servers managed by a session.

While working on that, I noticed that the technical debt entailed
by #184 and #185 started to be a bit too much. So, this work
aimed at making sure that we reduce the technical debt.

The general plan is to decouple the implementation of tests from
both the generic experiment and the generic session.

This diff addresses decoupling the specific experiment from the
generic experiment. To this end we introduce this ExperimentMeasurer
interface that now the generic experiment expects.

We cannot put this interface into the model package now, because of
a circular import. However, in the next commit will make it such that
this interface does not depend from session.

Once that is done, we can do #184 and #185 and finally #302.