[API CHANGE] Remove experiment/experiment.go and session/session.go
The code has been moved respectively to experiment.go and session.go.

In the process, the following API changes occurred:

1. ExperimentBuilder: Build => NewExperiment

2. Measurement: MarshalJSON removed

3. testslists API: marked as deprecated

The motivation of the first change is that NewExperiment is a more
idiomatic name, while Build is more Java style.

The motivation of the second and third change is that we are not
generating the mobile API directly from the toplevel engine API and
so there is no need to provide API that convert to ObjC/Java, and
instead we can write more idiomatic code.