[API CHANGE] remove session/session.go and experiment/experiment.go (#385)
As mentioned in https://github.com/ooni/probe-engine/issues/302, this split of the experiment and session was complicating applying further changes. It looked like the right moment to remove this source of technical debt and rationalise the implementation. Also, it's worth noting that we now have a strategy for producing mobile bindings that does not use the top-level "engine" API. So, a bunch of complexities that have been introduced, can be removed, making the "engine" API more idiomatic.

Summary of API changes:

1. log.Logger => model.Logger

2. ExperimentBuilder.{Build => NewExperiment}

3. engine.Callbacks => model.ExperimentCallbacks

4. added model.ExperimentMeasurer: this is the API between generic and specific experiment code

5. engine.Measurement merged with model.Measurement

6. added model.ExperimentSession: API exposed by session to specific experiment code

7. added model.ExperimentOrchestraClient: orchestra client exposed to specific experiment code

8. note that the testlists API may be deprecated

Closes https://github.com/ooni/probe-engine/issues/183

Closes https://github.com/ooni/probe-engine/issues/184