sniblocking: cache recent results
Avoids requesting several times in a row for the control SNI, which
is frankly a waste of time when checkoing multiple inputs.

While there, make sure we never use ps.ooni.io as the helper. A bunch
of places in tests were still using it, which is generally fine but
being consistent is most likely gonna be better.

Part of https://github.com/ooni/probe-engine/issues/309.