Write our own implementation of ndt7
The protocol is simple enough that we can afford that. This reduces the
coupling between our codebase and the official ndt7 client.

The only point of contact of the two codebases is the specification of
the messages that are sent by the server, which is wise to reuse.

Part of https://github.com/ooni/probe/issues/969.