oonimkall/runner: honour context w/ multiple inputs
As we do in ooni/probe-cli, don't interrupt a running experiment so we
collect more data. The main use case here is that one runs the whole test
list and has no way to interrupt OONI, so limit our actions to deal with
such use case for the time being.

Closes https://github.com/ooni/probe-engine/issues/351.