Replace the atomic package with atomicx (#414)
* Revert "oonitemplates: remove code causing crash on x86/arm64 (#360)"

This reverts commit 4e3bde35aea41bbdaf5e656924a97c034bace33d because it
makes sense to check whether this crash was actually caused by the alignment
instead https://github.com/ooni/probe-engine/issues/399.

Conflicts:
	internal/oonitemplates/oonitemplates.go
	internal/oonitemplates/oonitemplates_test.go

* Replace the atomic package with atomicx

The atomicx package is less efficient however it completely avoids
using atomic semantics on int64, which may lead to crash if the
variable we are accessing is not properly aligned.

See https://golang.org/pkg/sync/atomic/#pkg-note-BUG, which reads:

> On x86-32, the 64-bit functions use instructions unavailable
> before the Pentium MMX.
>
> On non-Linux ARM, the 64-bit functions use instructions
> unavailable before the ARMv6k core.
>
> On ARM, x86-32, and 32-bit MIPS, it is the caller's responsibility
> to arrange for 64-bit alignment of 64-bit words accessed atomically. The
> first word in a variable or in an allocated struct, array, or slice can
> be relied upon to be 64-bit aligned.

By using a separate package we are able to possibly even change the
way in which we implement the code, and surely we are able to shield
ourselves from this pitfalls by using a mutex. While in general it
may not be okay to always use a mutex, we mostly use atomic semantics
for counting. That is, we are not using it in performance critical
parts of the tree, hence this fix seems fine to me.

Related issue https://github.com/ooni/probe-engine/issues/399.