Kill the real_data_format_version annotation
This is part of the plan discussed with @FedericoCeratto
here https://github.com/ooni/probe-engine/issues/423

I am anticipating this part of the change so we don't release
a probe-cli that uses such annotation.

More changes to follow, after probe-engine 0.9.0, but I am not
going to guarantee 0.10.0 will happen before stable cli.