Only build for Android from the mobile-staging branch
This strategy is like the one adopted by Psiphon where they have
a staging branch into which they merge master. Such branch is tracking
the moments in time when the client is reasonably stable.

Likewise, we'll do the same for mobile. Our master branch is anyway
generally stable, because we're already doing CI and we're going
to improve upon that in the future.

Closes #376