Experiments in Go now use contex to report bytes usage
Luckily the mechanism for registering a byte counter with the context is
idempotent, so it does not matter a great deal if by mistake we, say, register
twice the experiment. In fact, there is one slot for the experiment and one
slot for the session, so it's difficult to make big mistakes in the code.

The worst that can happen is that we forget about registering in some cases
but double registrations are generally okay as long as the proper slot - for
the experiment or the session - is picked up.

Also, the places where we need to modify the context are few, so it seems
this is not going to be a very big burden. It's true that we need to remember
to use the proper dialer, but that's anyway a problem for all tests, to be
sure that we configure dialing correctly.

So, "burn the ships" like Cortez, and entirely remove the pre-existing
byte counting mechanism for Go experiments. We need to keep the callback
for byte counting, though, until we ditch MK experiments.

With this merged, https://github.com/ooni/probe-engine/issues/125 is
mostly done, except that we need to add a bunch of TODOs and follow up
actions after one more round of code inspection.