netx/{dialer,resolver}: actually add unit tests
They give us confidence that we can successfully log in case of
failure. This is a side effect and I still don't want to make sure
that the output does not change over time. Yet, it's important to
know that the code can handle the format strings we use.

Part of https://github.com/ooni/probe-engine/issues/359