Implement drop-in API-and-ABI-compatible MK replacement (#492)
This PR implements a drop-in API-and-ABI-compatible replacement for MK that can be used by FFI clients. The only FFI client I know is the soon-to-be-deprecated MK binary. I am not going to invest much time in automatically shipping the library we can create from this code. I am happy enough that we run tests and we have nearly full coverage. The only code bit without coverage is ~10 lines of obvious C code that just adds/removes the `const` qualifier.

With this functionality in, I am going to call https://github.com/measurement-kit/measurement-kit/issues/1913 done. Of course, we still need to rewrite everything in Go, but all the use cases covered by MK are nonetheless now fully covered. This is a precondition for https://github.com/measurement-kit/measurement-kit/issues/1921: if people need a C library, we can do the extra mile and build it for them.