Rework code that manages proxying
1. the existing approach was weird in that we were proxying only
some connections and not others. I remember this made some sense in
the past, but I don't fully remember now. It seems to me it's also
weird to not honour specific user settings. So, I have changed it to
honour the `--proxy` setting in full, also for measurements.

2. this allowed me to simplify the code a little bit.

3. but at the same time it seems I need to implement proxying at the
dialer level to be able to see all the events.

4. make sure that ndt7 and dash can be proxied over Tor.

Part of https://github.com/ooni/probe-engine/issues/506