ndt7 and dash can run over psiphon (#532)
The design is that we have a tunnel managed by the session. We may ask the
session to possibly start a tunnel, provided that one is configured.

If no tunnel is configured, nothing actually happens. Otherwise, we get
a proxy (similar to when `--proxy` is used) that we use.

The psiphon experiment now needs refactoring to use this tunnel.

Part of https://github.com/ooni/probe-engine/issues/506.