netx/httptransport: stop using http/httptrace
Using http/httptrace leads to tricky code because of the context and the
fact we may mistakenly register more than once.

If we want, we have network level data. So we are going to use network
level data to evaluate performance instead.

Part of https://github.com/ooni/probe-engine/issues/543