netx/httptransport: reintroduce http/httptrace
In https://github.com/ooni/probe-engine/pull/556 I was optimistic. It seems
it's difficult to generate a HAR w/o using http/httptrace.

Thus, reimplement measuring using http/httptrace. This time, however, do that
in a separate round tripper, to be more tidy. Also, structure the code such
that we really are not sharing the context in case of nested requests. I think
doing that increase the overall robustness, since context is a bit fragile.

Part of https://github.com/ooni/probe-engine/issues/543