Implement the tor tunnel (#585)
This implementation currently attempts to use an external tor
command and fails if not found. However, this is quite a giant
leap towards being able to use Tor in probe-engine.

You can optionally specify a specific tor command.It is also
possible to pass options to the tor command. This has allowed
me, e.g., to run DASH over obfs4+tor.

We're missing saving the logs. We're missing embedding Tor
where possible. But we can already do, though, urlgetter, dash, and
ndt7 over tor, like we do for Psiphon. This is nice because now it
means we can run performance tests also over tor.

This is part of https://github.com/ooni/ooni.org/issues/319

Closes https://github.com/ooni/probe-engine/issues/587