netx: implement self censorship mechanism
This is simpler than jafar. It has the advantage that it does not
require Linux to work. Basically, you need to compile with specific
compile tags to enable this (`-tags selfcensor`).

When you compile with such tags, MINIOONI_SELFCENSOR_SPEC is an
environment variable where you can set a specific JSON specification
describing how you'd like to self censor.

Note that, by default, the self censoring code is not enabled, as
you need to specify a tag for that. This means that you really need
to want to compile `miniooni` in this specific configuration.

We can censor the system resolver, TCP dials, and generally packets
that contain a specific fingerprint inside them.

Interestingly, I need this functionality for https://github.com/ooni/probe/issues/887
where I need to perform quick experiments and see what works.