netx/dialer/proxy.go: allow extemporary proxy
This diff allows us to use the context in order to setup an extemporary
proxy that is only valid within the context of the context.

While working on https://github.com/ooni/probe/issues/887, I have
determined that this is beneficial to switching over to using a
psiphon tunnel, where everything else fails.

While this is a bit of a hack, and I was not convinced about using the
context previously, not using the context seems to required lots of
burden because we need to setup and manage an alterntive HTTP client.