[API CHANGE] session: significantly simplify initialization
This diff significantly simplifies the way in which the session
is initialized. Now all options are set as part of the session
configuration and then everything follows nicely(TM).

This means that we don't need to worry when the session is alive
whether the programmer has changed, e.g., the collector URL.

Yak shaving as part of https://github.com/ooni/probe/issues/887.