libminiooni: allow starting a tunnel from command line
The use case is that we're blocked and then we start tor with, say, the
meek pluggable transport, and we can still reach the backends.

Manual runs start giving us some confidence that in such scenario we
can use tor to bypass censorship. To automate more this use case, sadly,
we need to ship tor and pluggable transports along with the desktop
app, therefore it's maybe the responsibility of the desktop app to
provide us with the required command line arguments.

How this would work is still unclear, yet the fact that we can submit
with lots of selfcensor censorship is at least reassuring.

Part of https://github.com/ooni/probe/issues/887.