session: tighten tunnel rules
For now, we cannot avoid the issue that a session only manages
a single tunnel or proxy, which is mostly sane BTW.

However, let's make sure that when we request for a differently
named tunnel, we fail. Otherwise, the psiphon experiment will
wrongly report success for psiphon when we are actually running
a session-wide tor instance forced from the command line.

Part of https://github.com/ooni/probe/issues/887