urlgetter: introduce the Multi urlgetter
With the Multi urlgetter you can perform operations in parallel with a
specific degree of parallelism. You get results back as they arrive via
a channel that is returned on output.

We'll use this functionality pretty much soon to rewrite the Telegram
experiment in terms of the urlgetter experiment.

Reference issue: https://github.com/ooni/probe-engine/issues/646