urlgetter: expose the operation that failed
This is very useful to quickly understand what actually went wrong. It is
a facet of the next design that was never really used.

Also, acknowledge that, in some cases, there isn't a real failed operation
because the error happens before we enter into our network stack.

This happens, for example, when the context we're using is canceled.

To address this potential issue, introduce and use a `"top_level"` operation
that describes this specific situation.

This work is part the telegram rewrite: https://github.com/ooni/probe-engine/issues/646.

Also, with this diff landing we can close https://github.com/ooni/probe-engine/issues/311,
which mentioned oonimkall (the old implementation of netx) but is still
basically talking about the same concept we're implementing here.