session.go: create subdirectory with config.TempDir (#687)
This diff changes the way in which session.go manages config.TempDir. Before this diff, the caller needed to supply a random directory where to store temporary session info. Instead, with this diff applied, the code uses the directory supplied by the caller as the base directory in which to create a randomly-named directory for storing temporary session info. Because the session owns that directory, it can `rm -rf` it when its `Close` method is called. Thus, we can guarantee that we do not leave junk around as long as the caller behaves and calls `Close` when done.

While working on this diff, I also cooperated with @lorenzoPrimi to understand whether leaving config.TempDir empty works everywhere. It certainly works on Desktop, where Go picks up the system-specific definition of temporary directory. It works on iOS. But it does not work on Android where `/data/local/tmp` is not writeable by the OONI application.

Nonetheless, I have decided to change the API and allow the caller, if they wish to do so, to specify an empty temporary directory. This is handy in cases in which you are running on Desktop and you just want the system to do the right thing. It allowed me, for example, to cancel a bunch of lines of code from the tests, defaulting to `/tmp` on Unix and something else on Windows.

We should probably not change anything in the Desktop and mobile apps, since it's fine to have them decide what the base temporary directory used by the engine should be.

Closes https://github.com/ooni/probe-engine/issues/671.