urlgetter: set DNSCache to nil when empty
Much better from the point of view of data analysis than having a
single entry which is actually empty.

We're currently limited to a single entry into the DNSCache but this
should not limit our data format, hence we keep an array.

The goal for the future is to have more than one entry in Config.

Part of https://github.com/ooni/probe-engine/issues/656