rename: internal/jsonapi => internal/httpx
To merge orchestra and probeservices, I also need to merge the two
supporting packages: jsonapi and fetch.

Specifically, we need to make fetch as cool as jsonapi and support also
cloudfronted and proxying via Tor/psiphon etc.

To this end, rename jsonapi to httpx, and start renaming the methods
such that it's clear that they deal with JSON input/output.

The next step will be to merge fetch with httpx.

Part of https://github.com/ooni/probe-engine/issues/651.