Merge internal/fetch with internal/httpx (#694)
The API needs to change slightly. Rather than fetching a full URL, we
split it into a base URL and a URL path.

Apparently, this is easily doable everywhere. I will need to change the
way in which we auto-generate assets.go though.

A nice side effect of this diff is that I can remove the urlpath package,
which should have been used in resources, bit was not.

While authoring this diff, I noticed that (1) there are too many cases
in which we take a pointer to the httpx.Client and we shan't, (2) we
should improve tests in httpx and resources to be more precise. Hence,
after this diff, I'll take a little detour and work on that.

Part of https://github.com/ooni/probe-engine/issues/651.