internal/orchestra: rewrite to have same constructor as probeservices
This should be one of the last steps. The orchestra client and the probeservices
client now have basically the same fields and the same constructor.

This means that it should be possible, in a follow-up step, to merge them.

In this diff, I am slightly reducing the code coverage. This will be amended in
a follow-up PR, once I have merged orchestra into probeservices.

Part of https://github.com/ooni/probe-engine/issues/651