probeservices: sketch out new API to access test lists (#702)
I want to replace the old test-list/urls API with this new API. The main
reason for doing so is that the new API uses the OrchestraClient instance
return by the session. The nice property of such client is that it will
use the endpoint we have settled to use when doing the selection of the best
probe service. Therefore, this would allow the apps to use either the fastest
service or the right circumvention method.

As part of merging this diff, I shall also create an issue about making sure
all users of the old API are migrated to the new API, such that we can remove
the old API. For now, the old API is just marked as deprecated.

Part of https://github.com/ooni/probe-engine/issues/651.