urlgetter: implement FailOnHTTPError
In some cases, we want to fail if we see an HTTP status code >= 400.

This is certainly the case when we're doing plain text HTTP.

Both the Telegram implementation and WhatsApp can benefit from that
functionality, when connecting to the web interface.

Thus, I've factored the code out of the Telegram implementation
and I have moved it inside the urlgetter package.

Part of https://github.com/ooni/probe-engine/issues/55