sniblocking: rewrite using urlgetter
A more substantial rewrite would have been possible. But let's proceed
piecemeal and, for now, just stop depending on the old netx.

This was the objective and we achieved it.

When we'll tackle the task of improving testing, we'll also refactor
the tests and see if we can also simplify sniblocking further.

Part of https://github.com/ooni/probe-engine/issues/684