sniblocking: rewrite using urlgetter (#744)
* sniblocking: rewrite using urlgetter

A more substantial rewrite would have been possible. But let's proceed
piecemeal and, for now, just stop depending on the old netx.

This was the objective and we achieved it.

When we'll tackle the task of improving testing, we'll also refactor
the tests and see if we can also simplify sniblocking further.

Part of https://github.com/ooni/probe-engine/issues/684

* sniblocking: further simplify using more urlgetter goodies

* sniblocking: add missing THAddress spotted during code review

This is part of https://github.com/ooni/probe-engine/issues/684

* sniblocking: start improving the quality of unit tests

Part of https://github.com/ooni/probe-engine/issues/684

* sniblocking: clarify what happens and add one more test

This gives us confidence that we obtain the right result even when we
are failing before even starting the experiment.

Part of https://github.com/ooni/probe-engine/issues/684.

* sniblocking: interrupted is a more proper error here

When the experiment was written, we didn't have this error. Now that we
have it, we should use it, because it's more proper.

Part of https://github.com/ooni/probe-engine/issues/684

* sniblocking: also specify the failed operation

This is more consistent.

Part of https://github.com/ooni/probe-engine/issues/684